Make data more realistic in terms of volume. Had to add some layers to get around Spring circular dependency problems.

+++ core/src/main/java/uk/ac/ucl/rits/inform/datasinks/emapstar/controllers/WaveformController.java
@@ -38,7 +38,7 @@ public class WaveformController {
     @Transactional
     public void processWaveform(WaveformMessage msg, Instant storedFrom) throws MessageIgnoredException {
         InterchangeValue<List<Double>> interchangeValue = msg.getNumericValues();
-        int samplingRate = msg.getSamplingRate();
+        long samplingRate = msg.getSamplingRate();
         if (!interchangeValue.isSave()) {
             throw new MessageIgnoredException("Updating/deleting waveform data is not supported");
         }
