+++ emap-interchange/src/main/java/uk/ac/ucl/rits/inform/interchange/visit_observations/WaveformMessage.java
@@ -29,7 +29,7 @@ public class WaveformMessage extends EmapOperationMessage {
     /**
      * Sampling rate in Hz.
      */
-    private int samplingRate;
+    private long samplingRate;
 
     /**
      * Numeric value.
